Rename nwpu_vhr10 to vhr10
