+++ torchgeo/datasets/__init__.py
@@ -74,7 +74,6 @@ from .loveda import LoveDA
 from .millionaid import MillionAID
 from .naip import NAIP
 from .nasa_marine_debris import NASAMarineDebris
-from .nwpu import VHR10
 from .openbuildings import OpenBuildings
 from .oscd import OSCD
 from .patternnet import PatternNet
@@ -104,6 +103,7 @@ from .utils import (
     unbind_samples,
 )
 from .vaihingen import Vaihingen2D
+from .vhr10 import VHR10
 from .xview import XView2
 from .zuericrop import ZueriCrop
